refactor: Merge infra_tools.py into sandbox_common.py
Fold snapshot and LSP tool builders from infra_tools.py into
sandbox_common.py where the aggregate build_sandbox_tools already
lives. Net: -1 file.
